app/pyproject.toml
[project]
name = "spdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "debugpy>=1.8.14",
    "folium>=0.19.5",
    "geopy>=2.4.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.0",
    "streamlit-folium>=0.25.0",
    "streamlit-modal>=0.1.2",
    "streamlit-option-menu>=0.4.0",
    "watchdog[watchmedo]>=6.0.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.14",
]
